refactor(core): split markdown parsing and chunking modules
